+++ assets/js/base.js
@@ -3,9 +3,6 @@
  * @author Michael Kühnel <m.kuehnel@micromata.de>
  */
 
-// JSHint settings
-/* jshint unused: false */
-
 /**
  * The jQuery object or a jQuery set containing on or more DOM elements.
  * @typedef {Object} jQuery
@@ -13,7 +10,7 @@
 
 /**
  * Namespace of the app.
- * This should be only variable the app should store in the global scope.
+ * This should be the only variable the app stores in the global scope.
  * @namespace kickstarter
  * @todo Rename the global variable throughout the whole file according to your needs.
  * @example
@@ -28,7 +25,7 @@ var kickstarter = window.kickstarter || {};
  * @namespace base
  * @memberof kickstarter
  */
-kickstarter.base = (function() {
+kickstarter.base = (function () {
 	'use strict';
 
 	/**
@@ -37,7 +34,7 @@ kickstarter.base = (function() {
 	 * @memberof kickstarter.base
 	 * @private
 	 */
-	var _fixViewportIssues = (function() {
+	(function () {
 		if (navigator.userAgent.match(/IEMobile\/10\.0/)) {
 			var msViewportStyle = document.createElement('style');
 			msViewportStyle.appendChild(
@@ -47,7 +44,7 @@ kickstarter.base = (function() {
 			);
 			document.querySelector('head').appendChild(msViewportStyle);
 		}
-	}());
+	})();
 
 	/**
 	 * Avoid `console` errors in browsers that lack a console.
@@ -55,9 +52,9 @@ kickstarter.base = (function() {
 	 * @memberof kickstarter.base
 	 * @private
 	 */
-	var _addConsoleMethods = (function() {
+	(function () {
 		var method;
-		var noop = function() {};
+		var noop = function () {};
 		var methods = [
 			'assert', 'clear', 'count', 'debug', 'dir', 'dirxml', 'error',
 			'exception', 'group', 'groupCollapsed', 'groupEnd', 'info', 'log',
@@ -75,6 +72,6 @@ kickstarter.base = (function() {
 				console[method] = noop;
 			}
 		}
-	}());
+	})();
 
-}());
+})();